move topology into its own file
